backend/core/pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["AlPetrovv <sanya10081999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-environ = "^0.11.2"
djoser = "^2.3.1"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
